Bit.ly short link support for files and directories - Also some UI tweaks and cleanup - Various code cleanups too

+++ src/main/java/onyx/components/search/indexer/IndexerJob.java
@@ -27,7 +27,9 @@
 package onyx.components.search.indexer;
 
 import com.google.common.collect.ImmutableSet;
+import onyx.components.search.SearchConfig;
 import onyx.components.search.SearchManager;
+import onyx.components.storage.AssetManager;
 import onyx.components.storage.ResourceManager;
 import onyx.entities.storage.aws.dynamodb.Resource;
 import org.apache.commons.lang3.time.DurationFormatUtils;
@@ -55,15 +57,23 @@ public final class IndexerJob implements Job {
             final JobExecutionContext context) throws JobExecutionException {
         final JobDataMap jobDataMap = context.getJobDetail().getJobDataMap();
 
+        final SearchConfig searchConfig =
+                (SearchConfig) jobDataMap.get(SearchConfig.class.getSimpleName());
         final ResourceManager resourceManager =
                 (ResourceManager) jobDataMap.get(ResourceManager.class.getSimpleName());
         final SearchManager searchManager =
                 (SearchManager) jobDataMap.get(SearchManager.class.getSimpleName());
+        final AssetManager assetManager =
+                (AssetManager) jobDataMap.get(AssetManager.class.getSimpleName());
 
         final long start = System.currentTimeMillis();
 
-        // First, nuke everything in the index.
-        searchManager.deleteIndex();
+        // Optionally, delete all documents in the index before a rebuild.
+        final boolean indexerRebuildDeleteIndexFirst =
+                searchConfig.getIndexerRebuildDeleteIndexFirst();
+        if (indexerRebuildDeleteIndexFirst) {
+            searchManager.deleteIndex();
+        }
 
         // Then, rebuild the entire index based on current resource data.
         int documentsIndexed = 0;
@@ -73,34 +83,46 @@ public final class IndexerJob implements Job {
 
             final Resource resource = resourceManager.getResourceAtPath(normalizedPath);
             if (resource != null) {
-                documentsIndexed += indexResource(resourceManager, searchManager, resource);
+                documentsIndexed += indexResource(resourceManager, searchManager, assetManager, resource);
             }
         }
 
         final long end = System.currentTimeMillis();
-        LOG.info("Successfully indexed {} resources in {}", documentsIndexed,
-                DurationFormatUtils.formatDurationHMS(end - start));
+        final String duration = DurationFormatUtils.formatDurationHMS(end - start);
+        LOG.info("Successfully indexed {} resources in {}", documentsIndexed, duration);
     }
 
     private int indexResource(
             final ResourceManager resourceManager,
             final SearchManager searchManager,
+            final AssetManager assetManager,
             final Resource resource) {
         int documentsIndexed = 0;
 
-        searchManager.addResourceToIndex(resource);
-        documentsIndexed++;
-        LOG.debug("Successfully indexed resource: {}", resource.getPath());
-
-        if (Resource.Type.DIRECTORY.equals(resource.getType())) {
+        if (Resource.Type.FILE.equals(resource.getType())) {
+            // If the file resource is in Dynamo, but does not resolve to a valid
+            // asset in S3 then log a warning and skip indexing of the file.
+            final boolean resourceExists = assetManager.resourceExists(resource);
+            if (!resourceExists) {
+                LOG.warn("Indexer skipping non-existent resource in S3: {}", resource.getPath());
+                return 0;
+            }
+        } else if (Resource.Type.DIRECTORY.equals(resource.getType())) {
+            // Recursively index the contents of the directory.
             final Set<Resource.Visibility> visibility =
                     ImmutableSet.of(Resource.Visibility.PUBLIC, Resource.Visibility.PRIVATE);
             final List<Resource> directoryContents =
                     resourceManager.listDirectory(resource, visibility, null);
             for (final Resource child : directoryContents) {
-                documentsIndexed += indexResource(resourceManager, searchManager, child); // Recursive!
+                // Recursive!
+                documentsIndexed += indexResource(resourceManager, searchManager,
+                        assetManager, child);
             }
         }
+
+        searchManager.addResourceToIndex(resource);
+        documentsIndexed++;
+        LOG.debug("Successfully indexed resource: {}", resource.getPath());
 
         return documentsIndexed;
     }
